Add readability example
As a matter of fact, this is a counter-example. Therefore, this
class is not meant to be executed and therefore it will not compile.
It's purpose is to illustrate a test case that does not respect
the readability principles.

It is pasted here almost exactly (some anonymization has taken place)
as I found it in a professional repository, a few years ago.